requirements.txt
requests
diskcache
gradio
loguru
prometheus-client
locust
smolagents
arize-phoenix 
opentelemetry-sdk 
opentelemetry-exporter-otlp 
openinference-instrumentation-smolagents